Added Dumbo - utilities for PostgreSQL handling
That will help to keep the shared code between Pongo and Emmett in one place.

Dumbo will be versioned separately, which should give possibility to use different versions in Pongo and Emmett. And have this peer dependency more stable.

Dumbo is treated for now as internal Pongo and Emmett dependency, so there are no plans (for now) to provide docs for it.