Changed _id type to TEXT
Changed _id type to Text. In PostgreSQL this should be almost equally the same indexable. Of course it'll take a bit more storage, but let's live with that for now.

We can in the future allow as Marten does to use different definition of collection table.

Changing from uuid to text will allow more sophisticated key strategies.

Most importantly it'll reuse the stream id as document id for Emmett projections.

Besides that added columns for future concepts:
- _version - for optimistic concurrency handling,
- _partition - for built-in PostgreSQL partitioning,
- _created, _modified - for timestamp metadata,
- metadata - for technical additional data,
- _archived - for soft deletion and also partitioning